pyproject.toml
[project]
name = "multi-tenant"
version = "0.1.0"
description = ""
authors = [
    {name = "Mohanraj",email = "mohanraj.mm1311@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.2)",
    "django-extensions (>=3.2.3)",
    "django-htmx (>=1.17.0)",
    "django-environ (>=0.11.2)",
    "pillow (>=12.1.1,<13.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "dnspython (>=2.8.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
flake8 = "^7.3.0"
isort = "^7.0.0"

